refactor(ios-tts-playback): split playback coordinator into focused files
- replace the 1,000+ line `iOS/KeyVox iOS/Core/TTS/TTSPlaybackCoordinator.swift` monolith with a dedicated `iOS/KeyVox iOS/Core/TTS/TTSPlaybackCoordinator/` directory
- keep `TTSPlaybackCoordinator.swift` focused on the main type, shared state, callbacks, and top-level playback/session entry points
- extract playback lifecycle and replay/session teardown into `TTSPlaybackCoordinator+Lifecycle.swift`
- extract stream scheduling, startup buffering, and buffer flush logic into `TTSPlaybackCoordinator+Scheduling.swift`
- extract playback metering helpers into `TTSPlaybackCoordinator+Metering.swift`
- extract playback progress, replay offset tracking, fast-mode safety state, and buffering helpers into `TTSPlaybackCoordinator+Progress.swift`
- move deterministic startup and runway policy rules into the pure `TTSPlaybackCoordinatorBufferingPolicy.swift` helper so buffering policy is separated from transport orchestration
- update `iOS/KeyVox iOS.xcodeproj/project.pbxproj` to add the new coordinator group and source memberships for the split files
- add `iOS/KeyVoxiOSTests/Core/TTS/TTSPlaybackCoordinatorBufferingPolicyTests.swift` to cover the extracted deterministic buffering policy behavior
